+++ rollup.config.js
@@ -50,6 +50,13 @@ const svelteSetup = {
         { find: "@data", replacement: path.resolve("src/data") },
         { find: "@components", replacement: path.resolve("src/ui/components") },
         { find: "@notifications", replacement: path.resolve("src/ui/notifications") },
+        { find: "@assets", replacement: path.resolve("src/ui/assets") },
+        { find: "@people", replacement: path.resolve("src/ui/people") },
+        { find: "@googleAdmin", replacement: path.resolve("src/ui/googleAdmin") },
+        { find: "@contracts", replacement: path.resolve("src/ui/contracts") },
+        { find: "@history", replacement: path.resolve("src/ui/history") },
+        { find: "@auth", replacement: path.resolve("src/ui/auth") },
+        { find: "@utils", replacement: path.resolve("src/ui/utils") },
         { find: "@ui", replacement: path.resolve("src/ui") }
       ]
     }),
