+++ src/ui/App.svelte
@@ -1,23 +1,23 @@
 <script lang="ts">
-  import Checkout from "@ui/Checkout.svelte";
-  import LookupAsset from "@ui/LookupAsset.svelte";
-  import Contracts from "@ui/Contracts.svelte";
-  import LookupStudent from "@ui/LookupStudent.svelte";
-  import History from "@ui/History.svelte";
+  import Checkout from "@assets/Checkout.svelte";
+  import LookupAsset from "@assets/LookupAsset.svelte";
+  import Contracts from "@contracts/Contracts.svelte";
+  import LookupStudent from "@people/students/LookupStudent.svelte";
+  import History from "@history/History.svelte";
   import Message from "@notifications/BulkMessageSender.svelte";
   import router from "page";
-  import LogIn from "@ui/LogIn.svelte";
+  import LogIn from "@auth/LogIn.svelte";
   import { loggedIn, user } from "@data/user";
-  import { l, parseMarkdown } from "@ui/util";
+  import { l, parseMarkdown } from "@utils/util";
   import { onMount, SvelteComponent } from "svelte";
   import { fade } from "svelte/transition";
-  import LookupStaff from "@ui/LookupStaff.svelte";
+  import LookupStaff from "@people/staff/LookupStaff.svelte";
   import Reports from "@ui/reports/Reports.svelte";
   import SISTest from "@ui/scheduling/SISTest.svelte";
   import ScheduleTester from "@ui/scheduling/ScheduleTester.svelte";
   import TicketBrowser from "@ui/tickets/TicketBrowser.svelte";
   import TicketNumberPage from "@ui/tickets/TicketNumberPage.svelte";
-  import Invoices from "@ui/Invoices.svelte";
+  import Invoices from "@contracts/Invoices.svelte";
   import Toast from "@components/Toast.svelte";
   import { toastStore } from "@components/toastStore";
   import TestMenu from "@ui/TestMenu.svelte";
