+++ src/ui/assets/ChromebookInfoDisplay.svelte
