+++ src/ui/auth/LogIn.svelte
@@ -2,7 +2,7 @@
   import netlifyIdentity from "netlify-identity-widget";
   import { onMount } from "svelte";
   import { user, loggedIn } from "@data/user";
-  import { l } from "./util";
+  import { l } from "@utils/util";
 
   onMount(() => {
     netlifyIdentity.init();
