+++ src/ui/tickets/editorComponents/TicketStudentAssignment.svelte
@@ -1,9 +1,9 @@
 <script lang="ts">
   import EditButton from "./EditButton.svelte";
-  import StudentTag from "../../StudentTag.svelte";
-  import NameDropdown from "../../NameDropdown.svelte";
+  import StudentTag from "@people/students/StudentTag.svelte";
+  import NameDropdown from "@people/components/NameDropdown.svelte";
   import { getStudent, studentsStore } from "@data/students";
-  import { studentName, validateStudent } from "@ui/validators";
+  import { studentName, validateStudent } from "@utils/validators";
   import type { Ticket } from "@data/tickets";
 
   export let ticket: Ticket;
